Fix regression with building empty dataloaders (#203)
One of the obsview changes seems to have broken downstream workflows, where empty DataLoaders used to be accepted but now they error. This reverts back to the previous behavior in this edge case as it's relied upon in some libraries for handling the case of test/train splits of 0% and 100% while being stable.

Found in https://github.com/SciML/ModelingToolkit.jl/actions/runs/13965380427/job/39094470015